refactor(twitchMonitor): split into focused modules to reduce complexity
File complexity was 146 (33 functions, 846 lines). Extracted into six
focused modules along domain boundaries, reducing the main file to
11 functions and 329 lines with a strict one-directional dependency chain.

New modules:
- twitchMonitorTypes.ts     — shared LiveState interface
- twitchMonitorEmbed.ts     — pure embed/template helpers
- twitchMonitorMultitwitch.ts — multitwitch URL building and footer updates
- twitchMonitorAnnouncements.ts — post/edit/delete Discord announcements
- twitchMonitorOffline.ts   — offline grace period timers
- twitchMonitorStartup.ts   — startup live-check sync
