+++ src/twitchMonitorStartup.ts
@@ -1,0 +1,139 @@
+import { discordClient } from './discordBot';
+import { isDiscordNotFoundError, tryDeleteDiscordMessage } from './discordUtils';
+import { getMonitorEnabled } from './monitorSettings';
+import { setStreamerLive, clearStreamerLive, DbStreamerFull } from './db';
+import { getStreams, TwitchStream } from './twitchApi';
+import { LiveState } from './twitchMonitorTypes';
+import { buildEmbed, fillTemplate, templateVars } from './twitchMonitorEmbed';
+import { updateMultitwitch } from './twitchMonitorMultitwitch';
+import { postAnnouncement } from './twitchMonitorAnnouncements';
+
+export async function tryEditStartupMessage(
+  liveStates: Map<string, LiveState>,
+  streamer: DbStreamerFull,
+  liveStream: TwitchStream,
+): Promise<boolean> {
+  if (!discordClient) return false;
+  if (!streamer.discord_channel_id || !streamer.discord_message_id) return false;
+  try {
+    const channel = await discordClient.channels.fetch(streamer.discord_channel_id);
+    if (!channel || !channel.isTextBased()) return false;
+    const vars = templateVars(streamer.name, liveStream);
+    const content = fillTemplate(streamer.group.live_message, vars);
+    const embed = buildEmbed(liveStream);
+    const message = await channel.messages.fetch(streamer.discord_message_id);
+    await message.edit({ content, embeds: [embed] });
+    liveStates.set(String(streamer.id), {
+      streamerId: streamer.id,
+      groupId: streamer.group.id,
+      group: streamer.group,
+      login: streamer.name.toLowerCase(),
+      messageId: streamer.discord_message_id,
+      channelId: streamer.discord_channel_id,
+      currentGame: liveStream.game_name,
+      title: liveStream.title,
+      currentStream: liveStream,
+      offlineTimer: null,
+    });
+    await setStreamerLive(streamer.id, streamer.discord_message_id, streamer.discord_channel_id, liveStream.game_name);
+    return true;
+  } catch (err) {
+    if (isDiscordNotFoundError(err)) return false;
+    console.error(`[TwitchMonitor] Failed to edit startup message for ${streamer.name}:`, err);
+    throw err;
+  }
+}
+
+export async function handleLiveStreamerOnStartup(
+  liveStates: Map<string, LiveState>,
+  streamer: DbStreamerFull,
+  liveStream: TwitchStream,
+  groupsWithChanges: Set<number>,
+): Promise<void> {
+  if (streamer.discord_message_id && streamer.discord_channel_id) {
+    if (!discordClient || !getMonitorEnabled()) {
+      // Disabled or no client: restore stored IDs into liveStates without touching Discord
+      liveStates.set(String(streamer.id), {
+        streamerId: streamer.id,
+        groupId: streamer.group.id,
+        group: streamer.group,
+        login: streamer.name.toLowerCase(),
+        messageId: streamer.discord_message_id,
+        channelId: streamer.discord_channel_id,
+        currentGame: liveStream.game_name,
+        title: liveStream.title,
+        currentStream: liveStream,
+        offlineTimer: null,
+      });
+      return;
+    }
+    try {
+      if (await tryEditStartupMessage(liveStates, streamer, liveStream)) {
+        groupsWithChanges.add(streamer.group.id);
+        return;
+      }
+    } catch {
+      // Edit failed (already logged) — skip postAnnouncement for this streamer
+      return;
+    }
+  }
+  await postAnnouncement(liveStates, streamer, liveStream);
+}
+
+export async function handleOfflineStreamerOnStartup(
+  streamer: DbStreamerFull,
+  groupsWithChanges: Set<number>,
+): Promise<void> {
+  // Stream ended while bot was offline — liveStates is empty at startup so
+  // deleteAnnouncement() would early-return without clearing DB state. Do it directly.
+  if (streamer.discord_channel_id && streamer.discord_message_id) {
+    try {
+      await tryDeleteDiscordMessage(streamer.discord_channel_id, streamer.discord_message_id);
+    } catch {
+      // Delete failed (already logged) — skip DB clear to avoid orphaning the announcement
+      return;
+    }
+  }
+  await clearStreamerLive(streamer.id);
+  groupsWithChanges.add(streamer.group.id);
+}
+
+export async function performStartupLiveCheck(
+  liveStates: Map<string, LiveState>,
+  loginToUserId: Map<string, string>,
+  streamersData: DbStreamerFull[],
+): Promise<void> {
+  const userIds = Array.from(loginToUserId.values());
+  if (userIds.length === 0) return;
+
+  let liveStreams: TwitchStream[] = [];
+  try {
+    liveStreams = await getStreams(userIds);
+  } catch (err) {
+    console.error('[TwitchMonitor] Startup live-check failed:', err);
+    return;
+  }
+
+  const liveByUserId = new Map(
+    liveStreams.filter((s) => s.type === 'live').map((s) => [s.user_id, s]),
+  );
+
+  const groupsWithChanges = new Set<number>();
+
+  for (const streamer of streamersData) {
+    const userId = loginToUserId.get(streamer.name.toLowerCase());
+    if (!userId) continue;
+
+    const liveStream = liveByUserId.get(userId);
+
+    if (liveStream) {
+      await handleLiveStreamerOnStartup(liveStates, streamer, liveStream, groupsWithChanges);
+    } else if (streamer.discord_message_id) {
+      await handleOfflineStreamerOnStartup(streamer, groupsWithChanges);
+    }
+  }
+
+  for (const gid of groupsWithChanges) {
+    await updateMultitwitch(gid, liveStates);
+  }
+}
